bug free with extra logs

+++ controller/repository.cpp
@@ -56,9 +56,9 @@ bool Repository::loadGroups(QString ownerId)
     bool returnValue = true;
     QSqlDatabase  db = QSqlDatabase::database(DbConnector::getConnectionName());
     QSqlQuery qry(db);
-    qry.prepare("select id, name, description from groups where ownerid = :id");
+    qry.prepare("select id, name, description from groups where ownerid = :ownerId");
 
-    qry.bindValue(":id", ownerId);
+    qry.bindValue(":ownerId", ownerId);
 
     if(qry.exec()){
             while(qry.next()){
@@ -90,15 +90,20 @@ bool Repository::loadCommercialGroupMembers(QString ownerId, QString groupId)
 
     qry.bindValue(":ownerId", ownerId);
     qry.bindValue(":groupId", groupId);
+    Logger::log("loadCommercialGroupMembers ownerId= "+QVariant(ownerId).toString()+" groupId= "+QVariant(groupId).toString());
 
     if(qry.exec()){
         while(qry.next()){
             QString commercialId = qry.value(0).toString();
             int index = (DbInterface::getInstance())->indexOfContactInContactList(commercialId, COMMERCIAL);
             if( index >= 0 )
+            {
+                Logger::log("group member index in contact list = "+QVariant(index).toString());
                 (DbInterface::getInstance())->appendNewMemberForGroup(index, groupId);
+            }
             else
             {
+                Logger::log("Can not find commercialId = "+ QVariant(commercialId).toString() + " s index in the contactList");
                 returnValue = false;
             }
         }
@@ -124,15 +129,20 @@ bool Repository::loadGeneralGroupMembers(QString ownerId, QString groupId)
 
     qry.bindValue(":ownerId", ownerId);
     qry.bindValue(":groupId", groupId);
-
+    Logger::log("loadGeneralGroupMembers ownerId= "+QVariant(ownerId).toString()+" groupId= "+QVariant(groupId).toString());
     if(qry.exec()){
         while(qry.next()){
             QString generalId = qry.value(0).toString();
             int index = (DbInterface::getInstance())->indexOfContactInContactList(generalId, GENERAL);
             if( index >= 0 )
+            {
+                Logger::log("group member index in contact list = "+QVariant(index).toString());
                 (DbInterface::getInstance())->appendNewMemberForGroup(index, groupId);
+            }
             else
             {
+                Logger::log("Can not find generalId = "+ QVariant(generalId).toString() + " s index in the contactList");
+
                 returnValue = false;
             }
         }
@@ -315,10 +325,14 @@ bool Repository::deleteCommercialGroupContactRelation(QString ownerId, QString g
     qry.bindValue(":groupId", groupId );
     qry.bindValue(":commercialId", commercialId);
 
+    Logger::log("deleteCommercialGroupContactRelation 1   ownerId: " + QVariant(ownerId).toString() + " groupId: " + QVariant(groupId).toString() + " commercialId: " + QVariant(commercialId).toString());
+
     if(!qry.exec()){
         Logger::log("deleting commercial contact error: "+ qry.lastError().text() );
         returnValue = false;
     }
+    Logger::log("deleteCommercialGroupContactRelation 2   ownerId: " + QVariant(ownerId).toString() + " groupId: " + QVariant(groupId).toString() + " commercialId: " + QVariant(commercialId).toString());
+
     return returnValue;
 }
 
@@ -335,10 +349,12 @@ bool Repository::deleteGeneralGroupContactRelation(QString ownerId, QString grou
     qry.bindValue(":groupId", groupId );
     qry.bindValue(":generalId", generalId);
 
+    Logger::log("deleteGeneralGroupContactRelation 1   ownerId: " + QVariant(ownerId).toString() + " groupId: " + QVariant(groupId).toString() + " generalId: " + QVariant(generalId).toString());
     if(!qry.exec()){
         Logger::log("deleting general contact error: "+ qry.lastError().text() );
         returnValue = false;
     }
+    Logger::log("deleteGeneralGroupContactRelation 2   ownerId: " + QVariant(ownerId).toString() + " groupId: " + QVariant(groupId).toString() + " generalId: " + QVariant(generalId).toString());
     return returnValue;
 }
 
@@ -348,13 +364,13 @@ bool Repository::deleteGroupFromDB(QString ownerId, QString groupId)
     QSqlDatabase  db = QSqlDatabase::database(DbConnector::getConnectionName());
     QSqlQuery qry(db);
 
-    qry.prepare("delete from Groups"
-                " where ownerid = :ownerId and groupId = :groupId");
+    qry.prepare("delete from Groups where ownerId = :ownerId and id = :groupId");
 
     qry.bindValue(":ownerId", ownerId);
     qry.bindValue(":groupId", groupId );
 
     if(!qry.exec()){
+        Logger::log("ownerId: " + QVariant(ownerId).toString() + " groupId: "+ QVariant(groupId).toString());
         Logger::log("deleting group error: "+ qry.lastError().text() );
         returnValue = false;
     }
